+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
 import type { NextPage } from 'next';
 import Head from 'next/head';
 import Image from 'next/image';
-import { About } from '../components/About/Hero';
+import { About } from '../components/About/myabout';
 import { Future } from '../components/Future/future';
 import { FuuterBar } from '../components/Pagefutter/Pagefutter';
 import { HeaderMain } from '../components/Pageheader/PageMain';
